Moved getPathById and getPathItem to Tree.
